requirements.txt
blessed==1.19.0
inquirer==2.8.0
python-editor==1.0.4
python-slugify==5.0.2
PyYAML==6.0
readchar==2.0.1
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
six==1.16.0
text-unidecode==1.3
wcwidth==0.2.5
